Documentation release.

+++ application/Locodio/Application/Command/Model/DeleteAssociation/DeleteAssociationHandler.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace App\Locodio\Application\Command\Model\DeleteAssociation;
 
+use App\Locodio\Application\Command\Model\ModelFinalChecker;
 use App\Locodio\Domain\Model\Model\AssociationRepository;
 
 class DeleteAssociationHandler
@@ -32,6 +33,10 @@ class DeleteAssociationHandler
     public function go(DeleteAssociation $command): bool
     {
         $relation = $this->associationRepo->getById($command->getId());
+        if (ModelFinalChecker::isFinalState($relation->getDomainModel()->getDocumentor())) {
+            return false;
+        }
+
         $this->associationRepo->delete($relation);
         return true;
     }
