+++ application/Locodio/Application/Command/Model/DeleteCommand/DeleteCommandHandler.php
@@ -2,6 +2,7 @@
 
 namespace App\Locodio\Application\Command\Model\DeleteCommand;
 
+use App\Locodio\Application\Command\Model\ModelFinalChecker;
 use App\Locodio\Domain\Model\Model\CommandRepository;
 
 class DeleteCommandHandler
@@ -21,6 +22,10 @@ class DeleteCommandHandler
     public function go(DeleteCommand $command): bool
     {
         $command = $this->commandRepo->getById($command->getId());
+        if (ModelFinalChecker::isFinalState($command->getDocumentor())) {
+            return false;
+        }
+
         $this->commandRepo->delete($command);
         return true;
     }
