+++ application/Locodio/Application/Command/Model/DeleteEnum/DeleteEnumHandler.php
@@ -2,6 +2,7 @@
 
 namespace App\Locodio\Application\Command\Model\DeleteEnum;
 
+use App\Locodio\Application\Command\Model\ModelFinalChecker;
 use App\Locodio\Domain\Model\Model\EnumOptionRepository;
 use App\Locodio\Domain\Model\Model\EnumRepository;
 
@@ -22,6 +23,10 @@ class DeleteEnumHandler
     public function go(DeleteEnum $command): bool
     {
         $enum = $this->enumRepo->getById($command->getId());
+        if (ModelFinalChecker::isFinalState($enum->getDocumentor())) {
+            return false;
+        }
+
         foreach ($enum->getOptions() as $option) {
             $this->enumOptionRepo->delete($option);
         }
